finished

+++ controllers/blogs.js
@@ -1,33 +1,57 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-
+const User = require('../models/user')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
     const body = request.body
 
     if (!body.title || !body.url) {
         return response.status(400).end()
     }
 
+    const user = request.user
+
     const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes || 0
+        likes: body.likes || 0,
+        user: user.id
     })
 
     const savedBlog = await blog.save()
+
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+    const user = request.user
+    if (!user) {
+        return response.status(401).json({ error: 'token missing or user not found' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+        return response.status(404).json({ error: 'blog not found' })
+    }
+
+    if (user.id.toString() === blog.user.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+        return response.status(204).end()
+    }
+    else
+        return response.status(400).json({ error: 'the blog does not belongs to you' })
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
